[Silver III] Title: 01타일, Time: 120 ms, Memory: 19588 KB -BaekjoonHub

+++ 백준/Silver/1904. 01타일/01타일.java
@@ -13,7 +13,7 @@ public class Main {
 		bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         int n = Integer.parseInt(br.readLine());
-		long[] dp = new long[n+1];
+		int[] dp = new int[n+1];
         dp[0] = dp[1] = 1;
         for (int i = 2; i <= n; i++) dp[i] = (dp[i-1] + dp[i-2]) % 15746;
